Revert "feat: support existing projects [sc-15038] (#640)"
This reverts commit 5e931943edee62cb786a36a9c01efe9a69a99c25.
